sheets-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic[email]
python-dotenv
PyJWT
pymongo
dnspython
user-agents
sib-api-v3-sdk
